Move ResponseCandidateStream to own module
